+++ NotePads/MenuConstants.java
@@ -1,0 +1,73 @@
+package NotePads;
+public interface MenuConstants {
+
+    String th_is = "This ";
+    String command = " command is yet to be implemented";
+    String aboutText2 = "About FNotepad!";
+    String fontChoose = "Choose a font";
+    String searchText = "Nothing to search for, use Find option of Edit Menu first !!!!";
+    String searchLine = "Enter Line Number:";
+    String letters = "Letters ";
+    String words = ", Words ";
+    String line = "       ||       Line ";
+    String column = ", Column ";
+    String countInit = "Letters 0, Words 0       ||       Ln 1, Col 1  ";
+    String defaultFileName = "Untitled";
+    String defaultApplicationName = "FNotepad";
+    String claf = "Change Look and Feel";
+    String matchCaseVar = "Match case";
+    String up = "Up";
+    String down = "Down";
+    String direction = "Direction";
+    String findText = "Find Next";
+    String replaceText = "Replace";
+    String replaceAllText = "Replace All";
+    String cancelText = "Cancel";
+
+
+
+    String fileText = "File";
+    String editText = "Edit";
+    String formatText = "Format";
+    String viewText = "View";
+    String helpText = "Help";
+
+    String fileNew = "New";
+    String fileOpen = "Open...";
+    String fileSave = "Save";
+    String fileSaveAs = "Save As...";
+    String filePageSetup = "Page Setup...";
+    String filePrint = "Print";
+    String fileExit = "Exit";
+
+    String editUndo = "Undo";
+    String editCut = "Cut";
+    String editCopy = "Copy";
+    String editPaste = "Paste";
+    String editDelete = "Delete";
+    String editFind = "Find...";
+    String editFindNext = "Find Next";
+    String editReplace = "Replace";
+    String editGoTo = "Go To...";
+    String editSelectAll = "Select All";
+    String editTimeDate = "Time/Date";
+
+    String formatWordWrap = "Word Wrap";
+    String formatFont = "Font...";
+    String formatForeground = "Set Text color...";
+    String formatBackground = "Set Pad color...";
+
+    String viewStatusBar = "Status Bar";
+
+    String helpHelpTopic = "Help Topic";
+    String helpAboutFNotepad = "About FNotepad";
+
+    String aboutText =
+            "<html><big>FNotepad</big><hr><hr>"
+                    + "<p align=right>From fantastic-octo-garbanzo!"
+                    + "<hr><p align=left>Compiled by OpenJDK15.<br><br>"
+                    + "<strong>Thank you for using FNotepad!</strong><br>"
+                    + "Please make an issue on<p align=center>"
+                    + "<hr><em><big>https://github.com/fantastic-octo-garbanzo/FNotepad</big></em><hr><html>";
+}
+
